postにuserを紐付け

+++ app/controllers/posts_controller.rb
@@ -6,8 +6,13 @@ class PostsController < ApplicationController
     @posts = Post.all
   end
 
+  def show
+    @post = Post.find(params[:id])
+  end
+
   def create
-    @post = Post.new(post_params)
+    @post = current_user.posts.build(post_params)
+    @posts = Post.all
     if @post.save
       redirect_to posts_path
     else
@@ -17,7 +22,7 @@ class PostsController < ApplicationController
 
   def destroy
     post = Post.find(params[:id])
-    post.destroy
+    post.destroy if post.user_id == current_user.id
     redirect_to root_path
   end
